Second and third Nimscodd missions to be completed upon occupation
"Gnomish Gnomish Pass" (which upon completion gives claims on the adjacent area, Dragondowns) is now completed as soon you occupy that area, rather than requiring you to conquer it in a peacedeal. Same goes for the following mission, "Continue the Push", which gives claims on the old Gnomish capital area.

The reason for this is:
- The completion rewards (i.e. claims) are most useful before the war is over, to reduce costs.
-Previously, it would be normal for a player to complete all three missions at the same time as they fully annexed Redscale. This way, a player will unlock them one by one.