dwarf regnal numbers
dwarf regnal numbers and a small change to a crimson deluge event. I also added Vertesk cores to the vertesker provinces in the empire owned by Gawed. I miss the automatic commit message generation bitbucket did.
Also I still haven't fixed the female heir thing. I talked with the main help person on the co-op and they just agreed that it was dumb. So that's still an issue. Nobody will actually get Halkraz or Ardeginn because nobody gets any female heirs. We need to do a test on it but I'm too busy with ravelianism.