Gor Burad minor stuff
Nerfing Basalt Byproducts, made it so that the player can't lock themselves by not completing Diverting the Lava before the disaster fires, as it now requires owning some provinces. Added a bunch of treasury deductions that were supposed to be there. Setting a variable to 0 when forming GB. Made it so that the Basalt Rage Army Infighting event will not happen during war.