Fix diggy menu allowing constructions
The decision to open the menu already checked for ongoing constructions,
but if you had a different valid province with no construction, you
that would let you open up the menu and start an expedition. In most
cases this was fine, but the dungeons need to not have any ongoing
construction (including promote settlement growth), otherwise the
monument doesn't get built.

Known bug: You can probably still be unlucky and start a construction
while your dungeon expedition is ongoing and block the monument that way